# studyindexer/app/static/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "StudyIndexer API",
    "description": "Document Processing and Semantic Search Service\n\nFeatures:\n* Document upload and processing (PDF, DOCX, TEXT, MARKDOWN)\n* Semantic search across documents\n* Course-specific document collections\n* Role-based access control\n* Document metadata management",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8000",
      "description": "Local development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "DocumentStatus": {
        "type": "string",
        "enum": ["pending", "processing", "completed", "failed"]
      },
      "DocumentType": {
        "type": "string",
        "enum": ["pdf", "text", "docx", "markdown"]
      },
      "CollectionType": {
        "type": "string",
        "enum": ["general", "course", "personal", "faq"]
      },
      "BaseResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "default": true
          },
          "message": {
            "type": "string",
            "nullable": true
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "ErrorDetail": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "details": {
            "type": "object",
            "nullable": true,
            "additionalProperties": true
          }
        }
      },
      "ErrorResponse": {
        "allOf": [
          {
            "$ref": "#/components/schemas/BaseResponse"
          },
          {
            "type": "object",
            "properties": {
              "success": {
                "type": "boolean",
                "default": false
              },
              "error": {
                "$ref": "#/components/schemas/ErrorDetail"
              }
            }
          }
        ]
      },
      "DocumentMetadata": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "author": {
            "type": "string"
          },
          "course_id": {
            "type": "string",
            "nullable": true
          },
          "document_type": {
            "$ref": "#/components/schemas/DocumentType"
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "collection": {
            "$ref": "#/components/schemas/CollectionType"
          },
          "custom_metadata": {
            "type": "object",
            "additionalProperties": true
          }
        }
      },
      "DocumentResponse": {
        "allOf": [
          {
            "$ref": "#/components/schemas/BaseResponse"
          },
          {
            "type": "object",
            "properties": {
              "document_id": {
                "type": "string",
                "description": "Unique document identifier"
              },
              "status": {
                "$ref": "#/components/schemas/DocumentStatus"
              },
              "message": {
                "type": "string",
                "nullable": true
              },
              "metadata": {
                "$ref": "#/components/schemas/DocumentMetadata"
              },
              "error": {
                "type": "string",
                "nullable": true
              }
            }
          }
        ]
      },
      "DocumentInfo": {
        "type": "object",
        "properties": {
          "document_id": {
            "type": "string",
            "description": "Unique document identifier"
          },
          "status": {
            "$ref": "#/components/schemas/DocumentStatus"
          },
          "metadata": {
            "$ref": "#/components/schemas/DocumentMetadata"
          },
          "error": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "DocumentListResponse": {
        "allOf": [
          {
            "$ref": "#/components/schemas/BaseResponse"
          },
          {
            "type": "object",
            "properties": {
              "documents": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/DocumentInfo"
                }
              },
              "total": {
                "type": "integer",
                "description": "Total number of documents"
              }
            }
          }
        ]
      },
...
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/api/v1/admin/documents": {
      "get": {
        "summary": "List all documents",
        "description": "Returns a list of all documents in the system (admin only)",
        "tags": ["admin"],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "description": "Maximum number of documents to return",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 10
            }
          },
          {
            "name": "offset",
            "in": "query",
            "description": "Number of documents to skip",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of documents",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentListResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/documents/my": {
      "get": {
        "summary": "List user documents",
        "description": "Returns a list of documents uploaded by the current user",
        "tags": ["documents"],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "description": "Maximum number of documents to return",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 10
            }
          },
          {
            "name": "offset",
            "in": "query",
            "description": "Number of documents to skip",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of user documents",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentListResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/documents/course/{course_id}": {
      "get": {
        "summary": "List course documents",
        "description": "Returns a list of documents for a specific course",
        "tags": ["documents"],
        "parameters": [
          {
            "name": "course_id",
            "in": "path",
            "description": "Course ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Maximum number of documents to return",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 10
            }
          },
          {
            "name": "offset",
            "in": "query",
            "description": "Number of documents to skip",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of course documents",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentListResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/documents/{document_id}": {
      "get": {
        "summary": "Get document",
        "description": "Returns details of a specific document",
        "tags": ["documents"],
        "parameters": [
          {
            "name": "document_id",
            "in": "path",
            "description": "Document ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Document details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Document not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  }
} 